Add Logging to Web Portal config
