+++ source/Octopus.Server.Client/Model/WebPortalConfiguration/WebPortalLoggingResource.cs
@@ -1,0 +1,9 @@
+﻿using Octopus.Client.Extensibility.Attributes;
+
+namespace Octopus.Client.Model.WebPortalConfiguration;
+
+public class WebPortalLoggingResource
+{
+    [Writeable]
+    public string[] TrustedProxies { get; set; }
+}
